restore user/project isolation to persistence
initializing the persistence object before the parent frame data
exists makes it leave off user+project info in the persistence
namespace. defaultConfig gets read on element create, when we
haven't yet had a chance to attach parent frame data to the element.
Quick fix to get us out of trouble is to initialize the recent*
persistence in attachedCallback along with the rest of the parent-
dependent state.