CA-2071: Show dashed segments on incomplete results
Note that this just mirrors the existing segmentation logic, which both
requires a customized version of Highcharts and fails to account for
incomplete data in the past (e.g., 2016-12-01 through 2017-01-01 with
monthly time unit will show data for only one day of January but a solid
line).

For future reference, IMO the right way to handle this is for
isIncomplete to take a "to" date and a unit, and determine whether that
"to" date is at the end of a unit boundary.

For avoiding a customized Highcharts, we could remove the last data
point from an incomplete segment, and create a new segment with just the
last two data points, linked to the prior segment and with different
styling. This works right now in line chart, but fails in stacked line
(though of course it seems to work in stacked line in the latest
Highcharts).