Fixes CA-1998: Cannot apply event filter to Custom & Autotrack Events
The root issue is custom events need to be queried with `$custom_event: <event_id>` for their properties.

The other issue is recent events don't store any data regarding whether the event is custom or not. So both the paths to getting properties for custom events need to be fixed.

This also fixes the custom event icon in recent events.