Use mp-calendar
Use the calendar (still need to update the mixpanel-common version in
package.json once the version with mp-calendar is published)

Also, change/fix some UX with the calendar/date inputs:
- If there’s a (start_date, end_date) range already selected, and you
select new_start_date which is > end_date, it’ll set the new range to be
(end_date, new_start_date). The inverse is true when you are selecting a
new_end_date.
- Picking a new_end_date that is the same as the current end_date will
cause the calendar dropdown to close.